Add the `ProjectCapture` action to `Focus`.
The `ProjectCapture` action allows users to work with
coproducts in a more intuitive way — the macro is
incomplete at the moment, since it assumes the end
encoding has inheritance.

It also adds the beginning of an `EmbeddedOptic` which
extends the DSL to be `open` so other parties can create
an `early-phase` keyword.